// frontend/src/api/generated/model/userGroupDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Nexmine API
 * OpenAPI spec version: v1
 */
import type { GroupMemberDto } from './groupMemberDto';

export interface UserGroupDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  adminUserId?: number | null;
  /** @nullable */
  adminUserName?: string | null;
  memberCount?: number;
  /** @nullable */
  members?: GroupMemberDto[] | null;
}